Updated both the navbar and the toasts to use react-bootstrap implementations
