+++ src/app/Contact.tsx
@@ -1,7 +1,6 @@
 'use client';
-// @ts-nocheck
-import Toast from 'bootstrap/js/dist/toast.js' 
-import { useRef } from 'react';
+import { useRef, useState } from 'react';
+import { Toast, ToastContainer } from 'react-bootstrap';
 
 /* 
   There's a piece of JS logic here to make a toast pop up when the e-mail address is pressed.
@@ -11,6 +10,7 @@ import { useRef } from 'react';
 
 export function Contact() {
   const toastRef = useRef(null)
+  const [show, setShow] = useState(false);
   return (
     <div>
       <div className="jumbotron jumbotron-fluid screenfull d-flex align-items-center" id="contact">
@@ -19,9 +19,16 @@ export function Contact() {
         <hr />
         <h2 className="fs-1">let&apos;s connect</h2>
         <p className="fs-3">if you want to reach out for absolutely any reason, my inbox is always open</p>
-        <p className="fs-3">e-mail me at: <span className="allen" id="e-mail" onClick={() => (
-          navigator.clipboard.writeText("allenschmerler at gmail dot com".replace(" at ", "@").replace(" dot ", ".")).then(() => Toast.getOrCreateInstance(toastRef.current!).show())
-        )}>allenschmerler at gmail dot com</span> <i className="bi bi-arrow-left mx-auto" id="mailArrow"></i></p>
+        <p className="fs-3">e-mail me at: 
+        <span className="allen" id="e-mail" onClick={() => {
+          navigator.clipboard.writeText("allenschmerler at gmail dot com".replace(" at ", "@").replace(" dot ", ".")).then(() => setShow(true))
+        }}
+          
+          
+        >allenschmerler at gmail dot com
+          </span> 
+          <i className="bi bi-arrow-left mx-auto" id="mailArrow"></i>
+        </p>
         <p className="fs-3">you can also use any of my social media links</p>
         <div className="fs-3 mb-3 d-flex " id="links">
           <div>
@@ -36,17 +43,28 @@ export function Contact() {
         </div>
       </div>
     </div>
-    <div className="toast-container position-fixed bottom-0 end-0 p-3">
-        <div id="liveToast" ref={toastRef} className="toast bg-dark text-light"  role="alert" aria-live="assertive" aria-atomic="true">
-          <div className="toast-header">
+    <div>
+    <ToastContainer
+          className="p-3 position-fixed"
+          position={'bottom-end'}
+          style={{ zIndex: 1 }}
+      >
+        <Toast className='bg-dark text-light' onClose={() => setShow(false)} show={show} delay={3000} autohide>
+          <Toast.Header>
+            <img
+              src="holder.js/20x20?text=%20"
+              className="rounded me-2"
+              alt=""
+            />
             <strong className="me-auto">Allen Schmerler</strong>
-            <button type="button" className="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
-          </div>
-          <div className="toast-body">
-            Copied the e-mail address to the clipboard succesfully!
-          </div>
-        </div>
-      </div>
+          </Toast.Header>
+          <Toast.Body>Copied the e-mail address to the clipboard succesfully!</Toast.Body>
+        </Toast>
+      </ToastContainer>
     </div>
+      
+        
+      </div>
+        
   );
 }
