+++ src/app/Notes.tsx
@@ -1,0 +1,15 @@
+export function Notes() {
+  return (
+    <div className="jumbotron jumbotron-fluid text-light mb-5" id="notes">
+      <div className="container">
+        <h1>notes</h1>
+        <hr />
+        <div id="notescontainer" className="d-flex flex-wrap justify-content-around">
+          {/*<!-- Notes -->*/}
+          <p>Normally, you&apos;d have a nice little section here with nice little post it notes. someone hates fun so this isn&apos;t a thing anymore. i&apos;m working on it!</p>
+        </div>
+      </div>
+
+    </div>
+  );
+}
